Changed size_t back to uint32.

+++ src/tier0/types.h
@@ -13,6 +13,6 @@ typedef unsigned int uint32_t;
 typedef long long int int64_t;
 typedef unsigned long long int uint64_t;
 
-typedef uint64_t size_t;
+typedef uint32_t size_t;
 
 #endif //__TYPES_H_INCLUDED
